+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Increase build speed of root configuration by only importing required Composites/Resources
 - Added ''UseEnvironment'' parameter to cater for RSOP for identical node names in different environments
 - Adding Home.md to wikiSource and correct casing.
+- Reset PSModulePath after MOF compilation of each node to avoid duplicate CIM class definition errors on the following nodes
 
 ### Fixed